+++ packages/nodes/tailwind.config.js
@@ -33,7 +33,7 @@ module.exports = {
       boxShadow: {
         main: '-2px 2px 0 0 rgba(123, 191, 165, 0.3)',
         input: 'inset -2px 2px 0px 0px rgba(123, 191, 165, 0.3)'
-      }
+      },
     },
   },
   plugins: [],
